Move testSet to lib
